+++ test/foundry/TestHelpers.sol
@@ -103,7 +103,7 @@ abstract contract TestHelpers is AssertionHelpers, TestParameters {
             maximumEntries: 200,
             maximumEntriesPerParticipant: 200,
             prizeValue: 1 ether,
-            minimumProfitBp: 500,
+            minimumProfitBp: uint16(500),
             feeTokenAddress: address(0),
             prizes: prizes,
             pricings: pricings
